default style for search component
